backend/requirements.txt
asgiref==3.11.1
Django==5.2.12
django-cors-headers==4.9.0
django-environ==0.13.0
django-phonenumber-field==8.4.0
djangorestframework==3.16.1
phonenumbers==9.0.26
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
